Disable SDL2 support for overlay renderers, merge into one file
