+++ src/graphics.c
@@ -68,8 +68,8 @@ static const struct renderer_data renderers[] =
 #endif
 #if defined(CONFIG_RENDER_SOFTSCALE)
   { "softscale", render_softscale_register },
-  //{ "overlay1", render_softscale_register },
-  //{ "overlay2", render_softscale_register },
+  { "overlay1", render_softscale_register },
+  { "overlay2", render_softscale_register },
 #endif
 #if defined(CONFIG_RENDER_GL_FIXED)
   { "opengl1", render_gl1_register },
